+++ lib/smart_form_fields.dart
@@ -3,6 +3,12 @@ library;
 
 export 'src/animation/smart_error_animation.dart' show SmartErrorAnimation;
 export 'src/form/smart_form.dart' show SmartForm, SmartFormState;
+export 'src/form/smart_api_errors.dart'
+    show
+        SmartApiErrorExtractor,
+        SmartApiErrorPayload,
+        SmartApiErrorResult,
+        SmartApiErrors;
 export 'src/form/smart_form_controller.dart' show SmartFormController;
 export 'src/form/smart_form_key.dart' show SmartFormKey;
 export 'src/form/smart_form_result.dart' show SmartFormResult;